Fix my_sleep_msecs argument type
The function prototype of my_sleep_msecs (in my.h) accepts an unsigned int
while its definition (in my.c) accepts an unsigned long int.

The function is only called from tui_sys_sleep which accepts an unsigned
long int, so the intended type seems to be unsigned long int.

Thus, the prototype is changed to reflect the intended type.