Remove MY_C macro
In my.h there is a check for the definition of the MY_C variable
and the function prototypes are declared only if it is not defined.

Every c file which includes my.h except my.c does not define MY_C,
so the function prototypes are declared correctly.

So the purpose of MY_C is to make sure the function prototypes
are not included in my.c which does not make any sense and is
therefore removed.