Split functions into OS specific files
Many of the functions in my.c use behavior which unfortunately is OS
specific. So preprocessor macros are utilized to determine the OS
and run different chunks of code. This leads to extremely long functions
which are difficult to read as demonstrated in the following example.

return_type function_name(parameters) {
	#if defined(OS_n1)
		OS_n1_foo();
		Several more lines of OS_n1 code;
	#elif defined(OS_n2)
		OS_n2_bar();
		Several more lines of OS_n2 code;
	#else
		Several lines of fallback code for unsupported systems;
	#endif
}

A different approach is to split the functionality into OS specific
files and use the preprocessor macros to include the appropriate files.