Document VMIN and VTIME
VMIN means the number of characters that must be available
in the input queue for read to return. VTIME is the timeout
value between characters.

The first variable that is honoured is the count parameter of read
function call. migf1 calls read with MY_MAGIC_MAX_CHARS (18) so
if the input queue contains 18 or more characters, read will
immediately return outputting 18 characters.

If the input queue contains less than 18 characters we come
to the VMIN / VTIME couple. VTIME kicks only after the first byte
is read, so if there is no character in the input queue, the call
will block until the first character is read. Afterwards, read
will return either if VMIN characters are read or VTIME is exceeded.

So, in practice, because VMIN is rather large, read call will return
1 * 0.1 seconds after the first character is read and output the
one (or two - three if a "special" key is pressed) it read.